break derived apart to the four winds
