+++ pkg/arm/v4/derive.go
@@ -1,0 +1,15 @@
+package arm
+
+import (
+	"strings"
+
+	"github.com/openshift/openshift-azure/pkg/api"
+)
+
+type derivedType struct{}
+
+var derived = &derivedType{}
+
+func (derivedType) MasterLBCNamePrefix(cs *api.OpenShiftManagedCluster) string {
+	return strings.Split(cs.Properties.FQDN, ".")[0]
+}
