+++ pkg/arm/v4/derive_test.go
@@ -1,0 +1,16 @@
+package arm
+
+import (
+	"testing"
+
+	"github.com/openshift/openshift-azure/pkg/api"
+)
+
+func TestDerivedMasterLBCNamePrefix(t *testing.T) {
+	cs := api.OpenShiftManagedCluster{
+		Properties: api.Properties{FQDN: "bar.baz"},
+	}
+	if got := derived.MasterLBCNamePrefix(&cs); got != "bar" {
+		t.Errorf("derived.MasterLBCNamePrefix() = %v, want %v", got, "bar")
+	}
+}
